Analysis of Rainfall Data for Jaipur and Jodhpur Divisions (1980-2020).md
This project calculates the **Standardized Precipitation Index (SPI)** for 40 years of rainfall data in the semi-arid and arid regions of **Jaipur** and **Jodhpur**, Rajasthan. The SPI is calculated for multiple time scales (SPI-3, SPI-6, SPI-9, SPI-12, SPI-24, and SPI-36) to analyze drought trends and variability in rainfall patterns. The project aims to provide insights into long-term drought conditions and their impacts on these regions.